:recycle: refactor codes
