+++ src/pages/games/tetris/composables/useO.ts
@@ -1,7 +1,9 @@
 import { useNodes } from './useNodes';
+import { usePositions } from './usePositions';
 
 export function useO() {
   const { nodes, NODE, TOP_RESERVE } = useNodes();
+
   const defaultPosition = Object.freeze([
     [2, 3],
     [2, 4],
@@ -9,66 +11,9 @@ export function useO() {
     [3, 4],
   ]);
 
-  const position = ref([...defaultPosition]);
   const shape = ref<0>(0);
-  const nextPosition = computed(() =>
-    position.value.map(([rowIdx, colIdx]) => [rowIdx + 1, colIdx]),
-  );
-  const endPosition = computed(computeEndPosition);
 
-  function computeEndPosition() {
-    let isBlocked = false;
-    let x = 0;
-    while (!isBlocked) {
-      x++;
-      for (let i = 0; i < position.value.length; i++) {
-        const [rowIdx, colIdx] = position.value[i];
-        if (
-          nodes.value[rowIdx + x][colIdx] === NODE.BIRTH ||
-          nodes.value[rowIdx + x][colIdx] === NODE.VOID ||
-          nodes.value[rowIdx + x][colIdx] === NODE.O
-        )
-          continue;
-        isBlocked = true;
-        break;
-      }
-    }
-    return position.value.map(([rowIdx, colIdx]) => [rowIdx + x - 1, colIdx]);
-  }
-
-  function prepare() {
-    shape.value = 0;
-    position.value = [...defaultPosition];
-    position.value.forEach(([rowIdx, colIdx]) => (nodes.value[rowIdx][colIdx] = NODE.O));
-  }
-
-  function fall() {
-    let isBlocked = false;
-    for (let i = 0; i < nextPosition.value.length; i++) {
-      const [nextRowIdx, nextColIdx] = nextPosition.value[i];
-      if (
-        nodes.value[nextRowIdx][nextColIdx] === NODE.BIRTH ||
-        nodes.value[nextRowIdx][nextColIdx] === NODE.VOID ||
-        nodes.value[nextRowIdx][nextColIdx] === NODE.O
-      )
-        continue;
-      isBlocked = true;
-      break;
-    }
-    if (isBlocked) {
-      position.value.forEach(([rowIdx, colIdx]) => (nodes.value[rowIdx][colIdx] = NODE.FOSSIL_O));
-      return false;
-    }
-    position.value.forEach(([rowIdx, colIdx]) => {
-      nodes.value[rowIdx][colIdx] = rowIdx <= TOP_RESERVE - 1 ? NODE.BIRTH : NODE.VOID;
-    });
-    nextPosition.value.forEach(([nextRowIdx, nextColIdx]) => {
-      nodes.value[nextRowIdx][nextColIdx] = NODE.O;
-    });
-
-    position.value = nextPosition.value;
-    return true;
-  }
+  const { position, endPosition, prepare, fall } = usePositions([...defaultPosition], shape);
 
   function moveRight() {
     let willCollide = false;
